% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{cd2Model}
\alias{cd2Model}
\title{Fitting Cobb-Douglas Models}
\usage{
cd2Model(formula, data, response, x1, x2, time, constrained = TRUE,
  save.data = TRUE, correlation = NULL, ...)
}
\arguments{
\item{data}{a data fram in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually as character strings.}

\item{constrained}{a logical indicating whether the parameters are constrained}

\item{correlation}{an optional \code{\link{corStruct}} object describing the
within-group correlation structure. See the documentation of \code{\link{corClasses}}
for a description of the available \code{corStruct} classes.
If a grouping variable is to be used,
it must be specified in the \code{form} argument to the \code{corStruct} constructor.
Defaults to \code{NULL}, corresponding to uncorrelated errors.}

\item{formala}{a formual of the form \code{response ~ x1 + x2 + time}}
}
\value{
a CDEmodel object, which is an lm object with some additional attributes.
}
\description{
Fitting Cobb-Douglas Models
}
\examples{
US <- subset(Calvin, Country=="US")
cd2Model(response = iGDP, x1 = iK, x2 = iL, time = iYear, data=US)
cd2Model(iGDP ~ iK + iL + iYear, data=US)
cd2Model(response = "iGDP", x1="iK", x2="iL", time="iYear", data=US)

}
