% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cesModel2}
\alias{cesModel2}
\title{Fit CES model}
\usage{
cesModel2(countryAbbrev, energyType = "none", baseHistorical,
  data = loadData(countryAbbrev = countryAbbrev, baseHistorical =
  baseHistorical), prevModel = NULL, algorithms = c("PORT", "L-BFGS-B"),
  nest = "(kl)e", rho = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5, -0.75, -0.9,
  -0.99), rho1 = c(9, 2, 1, 0.43, 0.25, 0.1, -0.1, -0.5, -0.75, -0.9, -0.99),
  digits = 6, ...)
}
\arguments{
  \item{data}{a data frame, if you want to use resampled
  data.}

  \item{baseHistorical}{the base directory from which to
  load historical data.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{nest}{one of \code{"(kl)"}, \code{"(le)"}, or
  \code{"(ek)"}. Use \code{"(kl)"} if you want a fit
  without energy, regardless of which energyType is
  specified.}

  \item{energyType}{Use \code{"none"} to specify a CES fit
  without energy, but only if \code{nest != "(kl)"}.}

  \item{fittingToResampleData}{a logical.  If \code{FALSE},
  you should also supply a value for \code{origModel}
  because \code{origModel} will be used to obtain the
  starting point for a gradient search.}

  \item{origModel}{a model}

  \item{prevModel}{a model used to start gradient searches.
  Use \code{NULL} if you want to use the default start
  locations AND do a grid search in sigma.}

  \item{rho,rho1}{Default values for \code{rho} and
  \code{rho1} are a grid upon which searches will be made.
  Note that \code{rho = 0.25} and \code{rho1 = 0.25} are
  included. These are the default starting values for
  \code{rho} and \code{rho1}, so that we don't need to do a
  fit from the default values. \code{rho = 0.25}
  corresponds to \code{sigma = 0.8}.}
}
\value{
a list of models
}
\description{
This function fits a CES model to original or resampled data.
}

