% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconUK-doc.R
\docType{data}
\name{EconUK}
\alias{EconUK}
\title{UK Economic Data}
\format{A data frame with the following variables
\itemize{
\item \code{Year} Year
\item \code{iYear} Indexed year (starting from 0)
\item \code{iGDP}  Indexed GDP (scaled so value is 1 in 1960)
\item \code{iK}    Indexed capital stock (scaled so value is 1 in 1960)
\item \code{iL}    Indexed labor (scaled so value is 1 in 1960)
\item \code{iXp}   Indexed primary exergy (scaled so value is 1 in 1960)
\item \code{iXu}   Indexed useful exergy (scaled so value is 1 in 1960)
}}
\source{
This is a subset of the data used in
Heun, M.K.; Santos, J.; Brockway, P.E.; Pruim, R.; Domingos, T.; Sakai, M.
"From Theory to Econometrics to Energy Policy: Cautionary Tales for Policymaking Using Aggregate Production Functions".
\emph{Energies} 2017, 10, 203.
}
\usage{
EconUK
}
\description{
Macroeconomic data for the UK from 1960 through 2010.
}
\examples{
if (require(ggplot2)) {
  ggplot(data = EconUK) +
    geom_line(aes(y = iGDP, x = Year))
}
}
\references{
\url{doi://10.3390/en10020203}
}
\keyword{datasets}
