% Generated by roxygen2 (4.0.0): do not edit by hand
\name{genAllResampleData}
\alias{genAllResampleData}
\title{Generates all resample data for all countries, all model types, and all energy types}
\usage{
genAllResampleData(method = "wild", n = numResamples(), baseResample,
  baseHistorical, ...)
}
\arguments{
  \item{method}{one of \code{"resample"},
  \code{"residual"}, \code{"wild"}, or \code{"debug"}}

  \item{n}{the number of resamples you want to perform}

  \item{baseResample}{the relative path to the directory
  where you want to store results}

  \item{baseHistorical}{the relative path to the directory
  in which the historical data resides}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function generates curve fit coefficients and models
and stores them to disk. The data are stored in an
object called "resampleData" and an object called "resampleModels".
These objects will have the same name
when they are loaded back from disk.
}

