% Generated by roxygen2 (4.0.0): do not edit by hand
\name{genResampleData}
\alias{genResampleData}
\title{Generates resample data, for the base fit and resample fits}
\usage{
genResampleData(modelType = modelTypes, countryAbbrev = countryAbbrevs,
  energyType = "Q", factor = "K", method = "wild", n, clobber = TRUE,
  verbose = FALSE, baseResample, baseHistorical)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev}{a string representing the country
  for which you want to load resample data.}

  \item{energyType}{one of \code{"none"}, \code{"Q"} (for
  thermal energy), \code{"X"} (for exergy), or \code{"U"}
  (for useful work).}

  \item{factor}{one of \code{"K"} (for capital stock),
  \code{"L"} (for labor), \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{method}{one of \code{"resample"},
  \code{"residual"}, \code{"wild"}, or \code{"debug"}}

  \item{n}{the number of resamples you want to perform}

  \item{clobber}{whether you want to overwrite any previous
  results}

  \item{verbose}{whether you want lots of output}

  \item{baseResample}{the relative path to the directory
  where you want to store results}

  \item{baseHistorical}{the relative path to the directory
  in which the historical data resides}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function generates curve fit coefficients and models
and stores them to disk. The data are stored in an
object called "resampleData" and an object called "resampleModels".
These objects will have the same name
when they are loaded back from disk.
}

