% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{getData}
\alias{getData}
\title{Extract data associated with an object}
\usage{
getData(object, ...)
}
\arguments{
\item{object}{an object from which you want to extract data.}
}
\value{
data associated with \code{object}.
}
\description{
Extract data associated with an object
}
