% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{getHistory}
\alias{getHistory}
\title{Extract fitting history of a model}
\usage{
getHistory(model)
}
\arguments{
\item{model}{the model object from which you want to extract the \code{history}.}
}
\value{
the fitting history for a model.
}
\description{
This convenience function returns the fitting history for a model.
}
