% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{linexModel}
\alias{linexModel}
\title{Fit LINEX models}
\usage{
linexModel(formula, data, response, x1, x2, x3, time, save.data = TRUE,
  correlation = NULL)
}
\arguments{
\item{formula}{a formula of the form \code{response ~ x1 + x2 + x3 + time}}

\item{data}{a data frame in which \code{formula} is evaluated}

\item{response}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x1}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x2}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{x3}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{time}{instead of specifying a formula, expressions for
the components can be specified individually.}

\item{correlation}{an optional \code{\link[nlme:corStruct]{nlme::corStruct}} object describing the
within-group correlation structure. See the documentation of \code{\link[nlme:corClasses]{nlme::corClasses}}
for a description of the available \code{\link[nlme:corStruct]{nlme::corStruct}} classes.
If a grouping variable is to be used,
it must be specified in the \code{form} argument to the \code{\link[nlme:corStruct]{nlme::corStruct}} constructor.
Defaults to \code{NULL}, corresponding to uncorrelated errors.}
}
\description{
Fit LINEX models
}
\examples{
data(EconUK)
linexModel(iGDP ~ iK + iL + iXp + iYear, data = EconUK)
linexModel(response = "iGDP", x1="iK", x2="iL", x3="iXp", time="iYear", data = EconUK)
naturalCoef(linexModel(response = "iGDP", x1="iK", x2="iL", x3="iXp", time="iYear", data = EconUK))

}
