% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadAllResampleData}
\alias{loadAllResampleData}
\title{Loads all resample data, for the base fit and resample fits}
\usage{
loadAllResampleData(modelType, energyType = "none", factor,
  countryAbbrevsOrder = countryAbbrevs, archive = NULL,
  baseResample = NULL)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{energyType}{one of \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{factor}{one of \code{"K"} (for capital stock),
  \code{"L"} (for labor), \code{"Q"} (for thermal energy),
  \code{"X"} (for exergy), or \code{"U"} (for useful
  work).}

  \item{countryAbbrevsOrder}{a vector of country
  abbreviations in the order you want them to appear in the
  data frame.}

  \item{archive}{path to a .zip archive containing resample
  model data, if resample data are to be loaded from the
  archive}

  \item{base}{the relative path of the top-level directory
  containing the resample data, if the data are to be
  loaded from a directory or an archive.}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function returns a data frame that contains all of the resample model coefficients
}

