% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadData}
\alias{loadData}
\title{Loads historical economic data}
\usage{
loadData(countryAbbrev, baseHistorical)
}
\arguments{
  \item{countryAbbrev}{the country/countries for which you
  want to load data}

  \item{baseHistorical}{the relative path of the directory
  containing the historical data.}
}
\value{
a data frame containing historical economic data
}
\description{
This function returns a data frame containing historical economic data
}

