% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{naturalCoef}
\alias{naturalCoef}
\title{Natural coefficients of a model}
\usage{
naturalCoef(object, ...)
}
\arguments{
\item{object}{the model object from which you want to extract the natural coefficients.}

\item{...}{additional arguments}
}
\value{
the coefficients of the model on a "natural" scale.
}
\description{
A convenience function that returns the "natural" coefficients of a model object.
}
