% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitting.R
\name{residuals.CDEmodel}
\alias{residuals.CDEmodel}
\title{Extract CDEmodel Residuals}
\usage{
\method{residuals}{CDEmodel}(object, type = c("response", "pearson",
  "normalized"), ...)
}
\arguments{
\item{object}{a CDEmodel object}

\item{type}{an optional character string specifying the type of residuals to
be used. If "response", the "raw" residuals (observed - fitted) are used;
else, if "pearson", the standardized residuals (raw residuals divided by
the corresponding standard errors) are used; else, if "normalized", the
normalized residuals (standardized residuals pre-multiplied by the inverse
square-root factor of the estimated error correlation matrix) are used.
Partial matching of arguments is used, so only the first character needs to
be provided. Defaults to "response".}

\item{...}{some methods for this generic function require additional
arguments. None are used in this method.}
}
\description{
The residuals of the linear model are extracted.
}
\note{
This method is identical to \code{residuals.gls()} expect that the
"std" and "label" attributes are not retained in the residuals object.
}
